+++ codes/main_test.go
@@ -175,19 +175,19 @@ func TestCodes(t *testing.T) {
 	})
 
 	Convey("Create Atomic Swap Bid Creation Request", t, func() {
-		for _, code := range xdr.CreateASwapBidCreationRequestResultCodeAll {
+		for _, code := range xdr.CreateAtomicSwapBidRequestResultCodeAll {
 			checkCode(t, code)
 		}
 	})
 
 	Convey("Cancel atomic swap bid", t, func() {
-		for _, code := range xdr.CancelASwapBidResultCodeAll {
+		for _, code := range xdr.CancelAtomicSwapAskResultCodeAll {
 			checkCode(t, code)
 		}
 	})
 
 	Convey("Create Atomic Swap Request", t, func() {
-		for _, code := range xdr.CreateASwapRequestResultCodeAll {
+		for _, code := range xdr.CreateAtomicSwapAskRequestResultCodeAll {
 			checkCode(t, code)
 		}
 	})