+++ db2/core/main.go
@@ -80,10 +80,6 @@ type QInterface interface {
 	Balances() BalancesQI
 	//KeyValue - creates new KeyValue query helper
 	KeyValue() KeyValueQI
-	// AtomicSwapBid - creates new AtomicSwapBid query helper
-	AtomicSwapBid() AtomicSwapBidQI
-	// AtomicSwapQuoteAsset - creates new AtomicSwapQuoteAsset query helper
-	AtomicSwapQuoteAsset() AtomicSwapQuoteAssetQI
 
 	Trusts() *TrustQ
 	Offers() *OfferQ