+++ ingest2/operations/handler.go
@@ -88,10 +88,10 @@ func NewOperationsHandler(operationsStorage operationsStorage, participantEffect
 			xdr.OperationTypeCreateSaleRequest: &createSaleRequestOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},
-			xdr.OperationTypeCreateAswapBidRequest: &createAtomicSwapBidRequestOpHandler{
+			xdr.OperationTypeCreateAtomicSwapAskRequest: &createAtomicSwapAskRequestOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},
-			xdr.OperationTypeCreateAswapRequest: &createAtomicSwapRequestOpHandler{
+			xdr.OperationTypeCreateAtomicSwapBidRequest: &createAtomicSwapBidRequestOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},
 			xdr.OperationTypeCreateWithdrawalRequest: &createWithdrawRequestOpHandler{
@@ -110,7 +110,7 @@ func NewOperationsHandler(operationsStorage operationsStorage, participantEffect
 			xdr.OperationTypeCheckSaleState: &checkSaleStateOpHandler{
 				manageOfferOpHandler: manageOfferOpHandlerInst,
 			},
-			xdr.OperationTypeCancelAswapBid: &cancelAtomicSwapBidOpHandler{
+			xdr.OperationTypeCancelAtomicSwapAsk: &cancelAtomicSwapAskOpHandler{
 				effectsProvider: effectsBaseHandler,
 			},
 			xdr.OperationTypeManageInvoiceRequest: &deprecatedOpHandler{},